Consolidate docker execute
